LPS-198126 Move class SftpUploader to class GoogleMerchantBackgroundTaskExecutor, and remove SftpUploader from source-formatter-suppressions.xml
