+++ modules/apps/commerce/source-formatter-suppressions.xml
@@ -2,7 +2,6 @@
 
 <suppressions>
 	<checkstyle>
-		<suppress checks="CamelCaseNameCheck" files="commerce-google-merchant/src/main/java/com/liferay/commerce/google/merchant/internal/sftp/SftpUploader\.java" />
 		<suppress checks="ResourcePermissionCheck" files="commerce-service/src/main/java/com/liferay/commerce/internal/security/permission/resource/AccountEntryModelResourcePermission\.java" />
 		<suppress checks="VariableNameCheck" files="commerce-google-merchant/src/main/java/com/liferay/commerce/google/merchant/internal/jsch/FingerprintHostKeyRepository\.java" />
 	</checkstyle>
